% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache_data}
\alias{cache_data}
\title{Cache data obtained from the API.}
\usage{
cache_data(x, dataset_id, type = c("metadata", "dataset"),
  cache_dir = "~/salinasr-cache", ...)
}
\arguments{
\item{x}{Either a \code{metadata} object or a \code{\link[tibble:tibble]{tibble::tibble}}.}

\item{dataset_id}{\code{character}, ID used to identify dataset.}

\item{type}{\code{character}, \code{"metadata"} or \code{"dataset"}.}

\item{cache_dir}{\code{character}, path to read/write data.}

\item{...}{Additional arguments passed on to \code{\link[readr:write_rds]{readr::write_rds()}}.}
}
\value{
\code{readr::write_rds} returns \code{x} invisibly.
}
\description{
Convenience function to minimize redundant API calls and generally be a polite Internet
citizen.
}
