% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_html}
\alias{clean_html}
\title{Remove HTML tags from a string.}
\usage{
clean_html(x)
}
\arguments{
\item{x}{A \code{character} \code{vector}.}
}
\value{
\code{character}, \code{x} with HTML tags removed.
}
\description{
Remove HTML tags from a string.
}
\examples{
library(salinasr)

html <- "<html>
         <body>
         <p>This is a paragraph.</p>
         <p>This is another paragraph.</p>
         </body>
         </html>"

clean_html(html)

}
\seealso{
\code{\link[stringr:str_replace_all]{stringr::str_replace_all}}, \code{\link[stringr:str_squish]{stringr::str_squish()}}, \code{\link[stringr:str_trim]{stringr::str_trim()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
