% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gum}
\alias{gum}
\title{A wrapper around \code{\link[glue:glue]{glue::glue()}} to remove \code{glue} class.}
\usage{
gum(..., .envir = parent.frame())
}
\arguments{
\item{...}{\code{character}, expressions to format.}

\item{.envir}{\code{environment} in which to evaluate expressions.}
}
\value{
\code{character}
}
\description{
\code{glue} provides string interpolation similar to Python's F-strings, but returned values
inherit the \code{glue} class. This seems to cause unexpected results in certain environments,
e.g. \code{\link[base:cat]{base::cat()}}.
}
\examples{
library(salinasr)

my_name <- "Brendan"
my_lang <- "R"

gum("My name is {my_name} and I am using {my_lang}.")

}
\seealso{
\code{\link[glue:glue]{glue::glue()}}
\code{\link[glue:glue_data]{glue::glue_data()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
