% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_empty}
\alias{is_empty}
\title{Does an object contain anything?}
\usage{
is_empty(x, recursive = TRUE)
}
\arguments{
\item{x}{An R object, typically a \code{list} or \code{vector}.}

\item{recursive}{\code{logical}, whether to \code{unlist} \code{x} to check for contents at any depth.}
}
\value{
\code{logical}, whether \code{x} contains anything.
}
\description{
Does an object contain anything?
}
\examples{
library(salinasr)

vec <- NULL
vec
is_empty(vec)

nested_list <- list(a = list(b = list(c = list(d = NULL), e = NULL), f = NULL), g = NULL)
nested_list
is_empty(nested_list)
is_empty(nested_list, recursive = FALSE)

}
\seealso{
\code{\link[base:unlist]{base::unlist}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
