% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{is_metadata}
\alias{is_metadata}
\title{Test if an object is \code{metadata}.}
\usage{
is_metadata(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
\code{logical}, \code{TRUE} if the object inherits from the \code{metadata} class.
}
\description{
Test if an object is \code{metadata}.
}
\examples{
library(salinasr)

meta <- sal_get_metadata("bikeways")

is_metadata(meta)
  
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
