% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_if_named}
\alias{keep_if_named}
\title{Filter an object to only contained elements with desired \code{.names}.}
\usage{
keep_if_named(x, .names)
}
\arguments{
\item{x}{A named R object.}

\item{.names}{\code{character}, vector of names by which to filter \code{x}.}
}
\value{
Those elements of \code{x} whose name is in \code{.names}.
}
\description{
Filter an object to only contained elements with desired \code{.names}.
}
\examples{
library(salinasr)

ages <- c(Tom = 20, Mary = 19, Sue = 43, Joe = 62)
ages
keep_if_named(ages, c("Mary", "Sue"))

}
\seealso{
\code{\link[rlang:is_named]{rlang::is_named()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
