% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_describe_dataset.R
\name{sal_describe_dataset}
\alias{sal_describe_dataset}
\alias{sal_describe_dataset.tbl_df}
\alias{sal_describe_dataset.character}
\alias{sal_describe_dataset.metadata}
\title{Describe a dataset's variables.}
\usage{
sal_describe_dataset(x)

\method{sal_describe_dataset}{tbl_df}(x)

\method{sal_describe_dataset}{character}(x)

\method{sal_describe_dataset}{metadata}(x)
}
\arguments{
\item{dataset}{\code{\link[tibble:tibble]{tibble::tibble}} as obtained by \code{sal_get_dataset()}.}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble}} with columns:
\item{\bold{\code{name}} }{names of variables}
\item{\bold{\code{label}} }{(hopefully) helpful clarification of variables' names}
\item{\bold{\code{type}} }{original data type of variables}
}
\description{
Describe a dataset's variables.
}
\section{Methods (by class)}{
\itemize{
\item \code{tbl_df}: Method to obtain description from a \code{tibble::tibble}

\item \code{character}: Method to obtain description from a \code{character}.

\item \code{metadata}: Method to obtain description from a \code{metadata} object obtained by \code{sal_get_metadata()}.
}}

\examples{
library(salinasr)

sal_get_metadata("bikeways") \%>\% 
  sal_get_dataset() \%>\% 
  sal_describe_dataset()

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
