% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_get_dataset.R
\name{sal_describe_vars}
\alias{sal_describe_vars}
\title{Describe a dataset's variables.}
\usage{
sal_describe_vars(dataset)
}
\arguments{
\item{dataset}{\code{\link[tibble:tibble]{tibble::tibble}} as obtained by \code{sal_get_dataset()}.}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble}} with columns:
\item{\bold{\code{name}} }{names of variables}
\item{\bold{\code{label}} }{(hopefully) helpful clarification of variables' names}
\item{\bold{\code{type}} }{original data type of variables}
}
\description{
Describe a dataset's variables.
}
\examples{
library(salinasr)

sal_get_metadata("bikeways") \%>\% 
  sal_get_dataset() \%>\% 
  sal_describe_vars()
  
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
