% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_fetch_catalog.R
\name{sal_fetch_catalog}
\alias{sal_fetch_catalog}
\title{Fetch catalog of available data sets.}
\usage{

  sal_fetch_catalog(rss_url = "https://cityofsalinas.opendatasoft.com/api/v2/catalog/exports/rss")
}
\arguments{
\item{rss_url}{\code{character}, RSS feed URL. \cr
Default: \code{"https://cityofsalinas.opendatasoft.com/api/v2/catalog/exports/rss"}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble}} data frame with columns:
\item{\bold{\code{title}} }{\code{character}, data set title.}
\item{\bold{\code{dataset_id}} }{\code{character}, data set ID to use in API requests.}
\item{\bold{\code{pub_date}} }{\code{POSIXct}, date-time published.}
\item{\bold{\code{description}} }{\code{character}, data set description.}
\item{\bold{\code{link}} }{\code{character}, URL linking to dataset.}
}
\description{
Fetch catalog of available data sets.
}
\examples{
library(salinasr)
suppressPackageStartupMessages(library(tidyverse))

catalog <- sal_fetch_catalog()

catalog

catalog \%>\% glimpse()

grocery_stores <- catalog \%>\% 
  filter(grepl("grocery stores", description))
  
grocery_stores \%>\% pull(description)

grocery_stores \%>\% pull(dataset_id)

}
\references{
\url{https://cityofsalinas.opendatasoft.com/pages/homepage/}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
