% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_dataset.R
\name{sal_finalize_dataset}
\alias{sal_finalize_dataset}
\title{Enforce a \code{dataset}'s data types, as described in its \code{metadata}.}
\usage{
sal_finalize_dataset(tbl, metadata, cache_data = TRUE, ...)
}
\arguments{
\item{tbl}{\code{\link[tibble:tibble]{tibble::tibble}}}

\item{metadata}{A \code{metadata} object as obtained by \code{sal_get_metadata()}.}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble}}
}
\description{
Intended for internal use, but exported as it may prove useful.
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
