% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{sal_get_dataset}
\alias{sal_get_dataset}
\title{Import a dataset as a \code{tibble} or \code{sf} \code{tibble}.}
\usage{
sal_get_dataset(metadata, prioritize_cache = TRUE,
  cache_dir = "~/salinasr-cache", ...)
}
\arguments{
\item{metadata}{A \code{metadata} object as obtained by \code{sal_get_metadata()}.}

\item{prioritize_cache}{\code{logical}, whether to prioritize cached data and avoid
unnecessary API calls. \cr
Default: \code{TRUE}}

\item{cache_dir}{\code{character}, path to the directory in which data is cached. \cr
Defailt: \code{"~/salinasr-cache"}}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble}}, which will also be an \code{\link[sf:st_sf]{sf::st_sf}} object if
\code{metadata} refers to data marked as having \code{"geo"} features.
}
\description{
Import a dataset as a \code{tibble} or \code{sf} \code{tibble}.
}
\examples{
library(salinasr)
library(sf, quietly = TRUE)
library(ggmap, quietly = TRUE)
library(ggplot2)
library(leaflet)

bikeways_sf <- sal_get_metadata("bikeways") \%>\% 
  sal_get_dataset()

bikeways_sf \%>\% 
  ggplot() +
  geom_sf()

sal_base_map <- get_map("Salinas, CA", zoom = 11)

ggmap(sal_base_map) +
  geom_sf(data = bikeways_sf, color = "red", inherit.aes = FALSE) 

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
