% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sal_get_export_link}
\alias{sal_get_export_link}
\title{Obtain a URL pointing to a specific dataset in the desired file type.}
\usage{
sal_get_export_link(metadata, file_type = c("geojson", "json", "csv",
  "shp"))
}
\description{
Obtain a URL pointing to a specific dataset in the desired file type.
}
\examples{
library(salinasr)

meta <- sal_get_metadata("bikeways")

meta \%>\% 
  sal_get_export_link()

}
