% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_get_metadata.R
\name{sal_get_metadata}
\alias{sal_get_metadata}
\title{Fetch a data set's meta data.}
\usage{
sal_get_metadata(dataset_id)
}
\arguments{
\item{dataset_id}{\code{character}}
}
\value{
\code{dataset} object, with
}
\description{
Fetch a data set's meta data.
}
\examples{
library(salinasr)

meta <- sal_get_metadata("bikeways")

meta$description

meta$keyword

}
