% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_read_data.R
\name{sal_read_data}
\alias{sal_read_data}
\title{Read a data set into an R object}
\usage{
sal_read_data(metadata)
}
\arguments{
\item{metadata}{an object obtained by \code{sal_get_metadata()}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble}}, which will also be an \code{\link[sf:st_sf]{sf::st_sf}} object if
\code{metadata} refers to data marked as having \code{"geo"} features.
}
\description{
Read a data set into an R object
}
\examples{
library(salinasr)
library(ggplot2)

bikeways_sf <- sal_get_metadata("bikeways") \%>\% 
  sal_read_data()
  
bikeways_sf \%>\% 
  ggplot() +
  geom_sf()

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
