% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_fetch_catalog.R
\name{sal_research_catalog}
\alias{sal_research_catalog}
\title{Output catalog information in a format more conducive to reading.}
\usage{
sal_research_catalog(sal_catalog, n = 5L, width = getOption("width"), ...)
}
\arguments{
\item{sal_catalog}{\code{\link[tibble:tibble]{tibble::tibble}}, as returned by \code{sal_fetch_catalog}.}

\item{n}{\code{integer}, the number of items to return.}

\item{width}{\code{integer}, number of characters per line. \cr
Default: \code{getOption("width")}}

\item{...}{Additional arguments to be passed to other methods. Not currently used.}
}
\value{
an \code{invisible} \code{NULL})
}
\description{
This is intended to be helpful when specific search parameters are unknown
or simply for exploration of available data.
}
\examples{
library(salinasr)
suppressPackageStartupMessages(library(tidyverse))

catalog <- sal_fetch_catalog()

catalog \%>\% 
  sal_research_catalog()

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
