% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinasr-package.R
\docType{package}
\name{salinasr}
\alias{salinasr}
\alias{salinasr-package}
\alias{salinasr-package}
\title{salinasr: A framework to programmatically obtain data sets from the Salinas Open Data
Portal API, transform them into tidy data frames, and conduct analyses.}
\description{
salinasr: A framework to programmatically obtain data sets from the Salinas Open Data
Portal API, transform them into tidy data frames, and conduct analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://brendan-g-knapp.github.io/salinasr/}
  \item Report bugs at \url{https://github.com/brendan-g-knapp/salinasr/issues}
}

}
\author{
Brendan Knapp \href{mailto:syknapptic@gmail.com}{syknapptic@gmail.com}
}
