% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_read_data.R
\name{sal_describe_vars}
\alias{sal_describe_vars}
\title{Describe a data set's variables.}
\usage{
sal_describe_vars(dataset)
}
\arguments{
\item{dataset}{\code{\link[tibble:tibble]{tibble::tibble}} as obtained by \code{sal_read_data()}.}
}
\description{
Describe a data set's variables.
}
\examples{
library(salinasr)

sal_get_metadata("bikeways") \%>\% 
  sal_read_data() \%>\% 
  sal_describe_vars()
  
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
