% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_get_dataset.R
\name{sal_get_dataset}
\alias{sal_get_dataset}
\title{Get a dataset as a useful R object.}
\usage{
sal_get_dataset(metadata)
}
\arguments{
\item{metadata}{an object obtained by \code{sal_get_metadata()}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble}}, which will also be an \code{\link[sf:st_sf]{sf::st_sf}} object if
\code{metadata} refers to data marked as having \code{"geo"} features.
}
\description{
Get a dataset as a useful R object.
}
\examples{
library(salinasr)
library(sf, quietly = TRUE)
library(ggmap, quietly = TRUE)
library(ggplot2)
library(leaflet)

bikeways_sf <- sal_get_metadata("bikeways") \%>\% 
  sal_get_dataset()

bikeways_sf \%>\% 
  ggplot() +
  geom_sf()

sal_base_map <- get_map("Salinas, CA", zoom = 11)

ggmap(sal_base_map) +
  geom_sf(data = bikeways_sf, color = "red", inherit.aes = FALSE) 

# outputs interactive leaflet map:
bikeways_sf \%>\% 
  leaflet() \%>\% 
  addProviderTiles(leaflet::providers$OpenStreetMap, group = "Street") \%>\% 
  addProviderTiles(providers$Esri.WorldImagery, group = "Optical") \%>\% 
  addPolylines(label = ~ name,
               popup = paste0(
                 "<b>Name:</b> ", bikeways_sf$name, "<br>",
                 "<b>Jurisdiction:</b> ", bikeways_sf$juris, "<br>",
                 "<b>Start:</b> ", bikeways_sf$start, "<br>",
                 "<b>End</b>: ", bikeways_sf$end, "<br>",
                 "<b>Miles:</b> ", round(bikeways_sf$miles, 2), "<br>",
                 "<b>Feet:</b> ", round(bikeways_sf$feet, 2) 
                 )
               ) \%>\% 
  addLayersControl(baseGroups = c("Street", "Optical"),
                   options = layersControlOptions(collapsed = FALSE))
  
}
\references{
\href{https://r-spatial.github.io/sf/}{sf: Simple Features for R}
}
\seealso{
\code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
