% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_get_export_link.R
\name{sal_get_export_link}
\alias{sal_get_export_link}
\alias{sal_get_export_link.character}
\alias{sal_get_export_link.metadata}
\title{Obtain a URL pointing to a specific dataset in the desired file type.}
\usage{
sal_get_export_link(x, type = c("json", "csv", "geojson", "xls", "shp",
  "ov2"))

\method{sal_get_export_link}{character}(dataset_id, type = c("json",
  "csv", "geojson", "xls", "shp", "ov2"))

\method{sal_get_export_link}{metadata}(x, type = c("json", "csv",
  "geojson", "xls", "shp", "ov2"))
}
\description{
Obtain a URL pointing to a specific dataset in the desired file type.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Method to obtain a an export link from a \code{character} representing a \code{dataset_id}.

\item \code{metadata}: Method to obtain a an export link from a \code{metadata} object.
}}

\examples{
library(salinasr)

meta <- sal_get_metadata("bikeways")

meta \%>\% 
  sal_get_export_link()

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
