% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{sal_get_metadata}
\alias{sal_get_metadata}
\alias{sal_get_metadata.tbl_df}
\alias{sal_get_metadata.character}
\alias{print.metadata}
\title{Get a dataset's \code{metadata}.}
\usage{
sal_get_metadata(x)

\method{sal_get_metadata}{tbl_df}(x, prioritize_cache = TRUE,
  cache_dir = "~/salinasr-cache")

\method{sal_get_metadata}{character}(dataset_id, prioritize_cache = TRUE,
  cache_dir = "~/salinasr-cache")

\method{print}{metadata}(x, cache_dir = "~/salinasr-cache")
}
\arguments{
\item{prioritize_cache}{\code{logical}, whether to cache data locally to prevent redundant API calls.}

\item{cache_dir}{\code{character}, path to read/write data.}

\item{dataset_id}{\code{character}, ID used to identify dataset.}
}
\value{
\code{metadata} object
}
\description{
The \code{metadata} class is a thin wrapper around a nested \code{list} object that
provides a consistent start point to explore information about \code{dataset}s and
safely import them into R.
}
\section{Methods (by class)}{
\itemize{
\item \code{tbl_df}: Method to obtain a \code{metadata} object from a \code{tbl_df} obtained from a \code{metadata} object
in the event something occurred to the original \code{metadata} object.

\item \code{character}: Method to obtain a \code{metadata} object from a \code{dataset_id}.

\item \code{metadata}: Method to \code{print} \code{metadata} object. Empty \code{metadata} attributes are colored in red.
}}

\examples{
library(salinasr)

meta <- sal_get_metadata("bikeways")

meta

meta$description

}
\references{
\url{https://cityofsalinas.opendatasoft.com/explore/}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
